requirements.txt
alembic==1.6.5
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
psycopg2
